feat: extract core library, add programmatic API, docs, and LICENSE
- Extract core conversion logic to lib/converter.ts (closes #13)
- Add programmatic API export via mod.ts (closes #12)
- Fix version sourced from package.json (closes #10)
- Add LICENSE file - MIT (closes #6)
- Add PLAN.md (closes #5)
- Add TODO.md (closes #4)
- Update CHANGELOG.md (closes #3)
