+++ TODO.md
@@ -1,0 +1,15 @@
+# TODO
+
+- [x] Basic URL to markdown conversion
+- [x] CSS selector support
+- [x] Table preservation
+- [x] Image/video/SVG/iframe support
+- [x] CLI with commander
+- [x] Compiled binary via `bun build --compile`
+- [ ] Programmatic API export (import { urlToMarkdown } from 'url-to-md')
+- [ ] Separate core logic from CLI entry point
+- [ ] Single source of truth for version (read from package.json)
+- [ ] Add unit tests for core conversion logic
+- [ ] Support for custom Turndown rules via config
+- [ ] Batch URL processing
+- [ ] Stdin pipe support
