+++ index.ts
@@ -1,17 +1,19 @@
 #!/usr/bin/env bun
 import { Command } from 'commander';
-import TurndownService from 'turndown';
-import { load } from 'cheerio';
 import { mkdir } from 'node:fs/promises';
 import { dirname } from 'node:path';
 import { existsSync } from 'node:fs';
+import { urlToMarkdown } from './lib/converter';
+
+// Read version from package.json (single source of truth)
+const pkg = await Bun.file(new URL('./package.json', import.meta.url)).json();
 
 const program = new Command();
 
 program
   .name('url-to-md')
   .description('Convert a URL to markdown with advanced table support and media preservation')
-  .version('1.0.0')
+  .version(pkg.version)
   .argument('[url]', 'The URL to convert')
   .argument('[selector]', 'The CSS selector to use for the main content', "[role='main']")
   .argument('[output]', 'The output file', 'README.md')
@@ -30,140 +32,13 @@ program
       process.exit(1);
     }
 
-    // Validate URL
-    let parsedUrl: URL;
-    try {
-      parsedUrl = new URL(url);
-    } catch {
-      console.error('Error: Invalid URL');
-      process.exit(1);
-    }
-    if (!['http:', 'https:'].includes(parsedUrl.protocol)) {
-      console.error('Error: URL must use http or https protocol');
-      process.exit(1);
-    }
-
     try {
       console.log(`Fetching ${url}...`);
-      const controller = new AbortController();
-      const timeout = setTimeout(() => controller.abort(), 10000);
-      const response = await fetch(url, {
-        signal: controller.signal,
-        headers: {
-          'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36',
-        },
-      });
-      clearTimeout(timeout);
+      const result = await urlToMarkdown(url, { selector });
 
-      if (!response.ok) {
-        throw new Error(`HTTP error! status: ${response.status}`);
-      }
-
-      const html = await response.text();
-      const $ = load(html);
-
-      let content;
-      const selected = $(selector);
-
-      if (selected.length === 0) {
+      if (!result.selectorMatched) {
         console.warn(`Warning: No elements found with selector "${selector}". Using entire body.`);
-        content = $('body').html() || html;
-      } else {
-        content = selected.html() || '';
       }
-
-      const turndownService = new TurndownService({
-        headingStyle: 'atx',
-        hr: '---',
-        bulletListMarker: '-',
-        codeBlockStyle: 'fenced',
-        emDelimiter: '_',
-      });
-
-      // Add support for tables
-      turndownService.addRule('tableCell', {
-        filter: ['th', 'td'],
-        replacement: function (content, node) {
-          return content.trim().replace(/\n/g, ' ') + ' |';
-        },
-      });
-
-      turndownService.addRule('tableRow', {
-        filter: 'tr',
-        replacement: function (content, node) {
-          const cells = content.trim();
-          return cells ? '| ' + cells + '\n' : '';
-        },
-      });
-
-      turndownService.addRule('table', {
-        filter: 'table',
-        replacement: function (content, node) {
-          const rows = content.trim().split('\n').filter(row => row.trim());
-          if (rows.length === 0) return '';
-
-          const headerRow = rows[0];
-          const separator = headerRow
-            .split('|')
-            .filter(cell => cell.trim())
-            .map(() => '---')
-            .join(' | ');
-
-          return '\n\n' + rows[0] + '\n| ' + separator + ' |\n' + rows.slice(1).join('\n') + '\n\n';
-        },
-      });
-
-      // Preserve images with alt text and src
-      turndownService.addRule('images', {
-        filter: 'img',
-        replacement: function (content, node: any) {
-          const alt = node.getAttribute('alt') || '';
-          const src = node.getAttribute('src') || '';
-          const title = node.getAttribute('title');
-
-          if (!src) return '';
-
-          const titlePart = title ? ` "${title}"` : '';
-          return `![${alt}](${src}${titlePart})`;
-        },
-      });
-
-      // Preserve videos
-      turndownService.addRule('videos', {
-        filter: 'video',
-        replacement: function (content, node: any) {
-          const $node = $(node);
-          const src = node.getAttribute('src') || $node.find('source').attr('src') || '';
-          const poster = node.getAttribute('poster');
-
-          if (!src) return content;
-
-          let markdown = `\n\n**[Video](${src})**\n\n`;
-          if (poster) {
-            markdown += `[![Video Thumbnail](${poster})](${src})\n\n`;
-          }
-          return markdown;
-        },
-      });
-
-      // Preserve SVGs as-is
-      turndownService.addRule('svg', {
-        filter: 'svg',
-        replacement: function (content, node: any) {
-          return '\n\n```svg\n' + $.html(node) + '\n```\n\n';
-        },
-      });
-
-      // Preserve iframes (for embedded content)
-      turndownService.addRule('iframe', {
-        filter: 'iframe',
-        replacement: function (content, node: any) {
-          const src = node.getAttribute('src') || '';
-          return src ? `\n\n**[Embedded Content](${src})**\n\n` : '';
-        },
-      });
-
-      const markdown = turndownService.turndown(content);
 
       // Create output directory if path contains slashes
       if (output.includes('/')) {
@@ -174,11 +49,10 @@ program
         }
       }
 
-      await Bun.write(output, markdown);
+      await Bun.write(output, result.markdown);
 
       console.log(`✓ Successfully converted to markdown`);
       console.log(`✓ Output saved to: ${output}`);
-
     } catch (error) {
       console.error('Error:', error instanceof Error ? error.message : String(error));
       process.exit(1);
