+++ package.json
@@ -1,9 +1,12 @@
 {
   "name": "url-to-md",
   "version": "1.0.0",
-  "module": "index.ts",
+  "module": "mod.ts",
   "type": "module",
   "private": true,
+  "exports": {
+    ".": "./mod.ts"
+  },
   "bin": {
     "url-to-md": "./index.ts"
   },
